+++ app/services/queries/__init__.py
@@ -5,7 +5,7 @@ from .immunity_query_service import ImmunityQueryService
 from .models import (
     DpsBreakdownRow,
     DpsRow,
-    ImmunitySummaryRow,
+    ImmunityDisplayRow,
     TargetSummaryRow,
 )
 from .target_summary_query_service import TargetSummaryQueryService
@@ -14,8 +14,8 @@ __all__ = [
     "DpsBreakdownRow",
     "DpsRow",
     "DpsQueryService",
+    "ImmunityDisplayRow",
     "ImmunityQueryService",
-    "ImmunitySummaryRow",
     "TargetSummaryRow",
     "TargetSummaryQueryService",
 ]